% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coll.R
\name{new_coll}
\alias{new_coll}
\alias{new_coll,PostgreSQLConnection,sf,integer-method}
\alias{new_coll,PostgreSQLConnection,data.frame,integer-method}
\alias{new_coll,PostgreSQLConnection,sf,numeric-method}
\alias{new_coll,PostgreSQLConnection,data.frame,numeric-method}
\alias{new_coll,PostgreSQLConnection,sf,character-method}
\alias{new_coll,PostgreSQLConnection,data.frame,character-method}
\title{Insert new collections to database.}
\usage{
new_coll(db, sf, bulk, ...)

\S4method{new_coll}{PostgreSQLConnection,sf,integer}(db, sf, bulk, ...)

\S4method{new_coll}{PostgreSQLConnection,sf,numeric}(db, sf, bulk, ...)

\S4method{new_coll}{PostgreSQLConnection,sf,character}(db, sf, bulk, ...)
}
\arguments{
\item{db}{Connections to the database as \linkS4class{PostgreSQLConnection}.}

\item{sf}{An object of class \link[sf:sf]{sf} to be appended into the table
\strong{collections} (schema \strong{specimens}).}

\item{bulk}{Either an integer or a character value indicating the ID or the
name of the corresponding bulk, respectively. Note that you can insert
new records to only one bulk at once. If you use a character value, the
ID will be retrieved from the database (table \strong{projects}, schema
\strong{specimens}). If the name is not in database, it will be considered as
a new bulk and inserted accordingly.}

\item{...}{Further arguments passed among methods.}
}
\description{
A straight forward method to insert new collections in the specimens
database.
This will automatically create the respective specimens.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
