% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_list.R
\name{project_list}
\alias{project_list}
\alias{project_list.PostgreSQLConnection}
\alias{project_list,PostgreSQLConnection-method}
\title{Render a list of all projects' descriptions.}
\usage{
project_list(db, ...)

\method{project_list}{PostgreSQLConnection}(
  db,
  title = "List of Collections",
  output = "pdf_document",
  rmd_args = list(),
  ...
)
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed to \code{\link[=render_rmd]{render_rmd()}}.}

\item{title}{Character value with the title of the document.}

\item{output}{character value indicating the format of the output document.}

\item{rmd_args}{Named list of further arguments passed to \code{\link[=list2rmd_doc]{list2rmd_doc()}}.}
}
\description{
A document listing all project descriptions in an overview.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
