% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_specimens.R
\name{read_specimens}
\alias{read_specimens}
\alias{read_specimens.PostgreSQLConnection}
\alias{read_specimens,PostgreSQLConnection-method}
\title{Import Specimen Information.}
\usage{
read_specimens(db, ...)

\method{read_specimens}{PostgreSQLConnection}(
  db,
  adm,
  bulk,
  schema = "specimens",
  tax_schema = "plant_taxonomy",
  get_coords = TRUE,
  ...
)
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods (not yet used).}

\item{adm}{Connection to the database containing administrative units as
\linkS4class{PostgreSQLConnection}. If not provided, no administrative units
will be included in the output.}

\item{bulk}{Integer vector including the ID's of the requested bulks
(campaigns or projects).}

\item{schema}{Character value indicating the name of the schema containing
the specimens' information.}

\item{tax_schema}{Character value indicating the name of the schema containng
the associated taxonomic list.}

\item{get_coords}{Logical values indicating whether formatted coordinates
should be extracted from geometry or not.The default is
\code{'get_coords = TRUE'} but it may cause an error if some cells in the
geometry are empty.}
}
\value{
An S3 object of class \code{specimens}.
}
\description{
Import function for herbarium specimens stored in own database.
Information is arranged into three tables and formatted as an object of class
\linkS4class{specimens}.

For convenience the function will display the list of updates and prompt a
yes/no/cancel question. Here you should check the printed names.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
