% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\alias{release.specimens}
\alias{release,specimens-method}
\title{Release a table for an herbarium}
\usage{
release(x, ...)

\method{release}{specimens}(x, herb, trans, ...)
}
\arguments{
\item{x}{A \linkS4class{specimens} object retrieved by \code{\link[=read_specimens]{read_specimens()}}.}

\item{...}{Further arguments passed among methods (not in use).}

\item{herb}{A character value containig the code of the herbarium for the
release.}

\item{trans}{A list containing translations of variable names to the
requested template. The name of the elements represent an herbarium
(usually as code by the Index Herbariorum) and contain a database with
two columns: \strong{in} for the column name of the input object (after
applying \code{\link[=as_data.frame]{as_data.frame()}}), and the column \strong{out} with the names
respective column names in the output. For columns in the requested
output without representative in the \linkS4class{specimens} object, you may
indicate it with \code{NA} values at \strong{in}. If not provided, this function
will use a pre installed translator (check with
\code{specimensDB:::translator}).}
}
\value{
A data frame with columns sorted and named for release at a respective
herbarium.
}
\description{
Processing data frame obtained by \code{\link[=read_specimens]{read_specimens()}} to be released at a
respective herbarium.

Some herbaria have a template for storage of data.
}
