% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_description.R
\name{render_description}
\alias{render_description}
\alias{render_description.PostgreSQLConnection}
\alias{render_description,PostgreSQLConnection-method}
\title{Rendering bulk description.}
\usage{
render_description(db, ...)

\method{render_description}{PostgreSQLConnection}(db, bulk, output = "pdf_document", output_file, ...)
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further Arguments passed to \code{\link[=list2rmd_doc]{list2rmd_doc()}}.}

\item{bulk}{The identifies of the bulk (project) in the database.}

\item{output}{Character value or list with the output settings for the yaml
head.}

\item{output_file}{Character value with the name and path to the ouput file.
I is passed to \code{\link[=render_rmd]{render_rmd()}}.}
}
\description{
Producing a PDF file with the description of the bulk (project) using
\code{\link[=render_rmd]{render_rmd()}}.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
