% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spec.R
\name{split_spec}
\alias{split_spec}
\alias{split_spec.PostgreSQLConnection}
\alias{split_spec,PostgreSQLConnection-method}
\title{Split specimens}
\usage{
split_spec(db, ...)

\method{split_spec}{PostgreSQLConnection}(db, spec_id, add = 1, ...)
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments. Parameters added here will be used to set
values in the specimens table that are specific for duplicates. Such
values will be recycled in the case of multiple duplicates.}

\item{spec_id}{Integer value indicating the ID of the specimen that will be
split.}

\item{add}{Integer value indicating the number of duplicates that need to be
created.}
}
\description{
Single collections may be split into multiple specimens, either stored in the
same herbarium or distributed to different locations.
This function duplicates specimens belonging to the same collection.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
