% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_specimens.R
\name{split_specimens}
\alias{split_specimens}
\alias{split_specimens.PostgreSQLConnection}
\alias{split_specimens,PostgreSQLConnection-method}
\title{Split specimens}
\usage{
split_specimens(db, ...)

\method{split_specimens}{PostgreSQLConnection}(db, spec_id, add = 1, schema = "specimens", ...)
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments. Parameters added here will be used to set
values in the specimens table that are specific for duplicates. Such
values will be recycled in the case of multiple duplicates.}

\item{spec_id}{Integer vector with the IDs specimens that will be duplicated.
If one ID is not existing in the database, this function will retrieve an
error message.}

\item{add}{Integer vector indicating the number of duplicates that need to be
created. If the length is 1, the same amount of duplicates will be
created for every specimen. For specific number of duplicates, you have
to provide a vector with the same length as 'spec_id'. If the lenght is
neither 1 nor matching the length of 'spec_id', an error message will be
retrieved.}

\item{schema}{Character value indicating the name of the database schema
containing specimens.}
}
\description{
Single collections may be split into multiple specimens, either stored in the
same herbarium or distributed to different locations.
This function duplicates specimens belonging to the same collection.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
