% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_det.R
\name{update_det}
\alias{update_det}
\alias{update_det,PostgreSQLConnection,data.frame-method}
\title{Updating determinations of specimens.}
\usage{
update_det(db, df, ...)

\S4method{update_det}{PostgreSQLConnection,data.frame}(db, df, ...)
}
\arguments{
\item{db}{Connections to the database as \linkS4class{PostgreSQLConnection}.}

\item{df}{A data frame containing the information to be appended in the
determination history of the specimen. This data frame must contain
following columns: \strong{spec_id} (identifier of specimens),
\strong{taxon_usage_id} (identifier of the taxon usage name), \strong{taxonomy}
(name of the taxonomic list applied to the update), \strong{det} (name of the
person that determined the species), and \strong{det_date} (date of
determination as \link[=as.Date]{Date}).}

\item{...}{Further arguments passed among methods (not yet used).}
}
\value{
SQL commands will be executed.
}
\description{
Inserting new determinations to specimens stored in the database.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
