% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_envelopes.R
\name{write_envelopes}
\alias{write_envelopes}
\alias{write_envelopes.specimens}
\alias{write_envelopes,specimens-method}
\title{Writting labels for envelopes}
\usage{
write_envelopes(x, ...)

\method{write_envelopes}{specimens}(
  x,
  output_file,
  date_format = "\%d.\%m.\%Y",
  mar = c(l = 31, r = 31, t = 43, b = 148, u = "mm"),
  frame = FALSE,
  classoption = "a4paper",
  ...
)
}
\arguments{
\item{x}{A \linkS4class{specimens} object retrieved by \code{\link[=read_spec]{read_spec()}}.}

\item{...}{Further arguments passed to \code{\link[=write_rmd]{write_rmd()}}. It works only if
\code{'merge = TRUE'}.}

\item{output_file}{A character value indicating the name of the file. It may
include a relative path. This is passed to the homonymous parameter at
\code{\link[=render_rmd]{render_rmd()}}.}

\item{date_format}{A character value indicating the format used for the
collection date. It is passed to the function \code{\link[=format.Date]{format.Date()}}.}

\item{mar}{A named vector indicating the margins. The mandatory names are
\strong{l} (left), \strong{r} (right), \strong{t} (top), \strong{b} (bottom), and
\strong{u} (units).}

\item{frame}{A logical value indicating whether a frame around the label
should be drawn or not.}

\item{classoption}{A character value to be inserted as 'classoption' in the
yaml head for the Rmarkdown document.}
}
\value{
An invisible \linkS4class{rmd_doc} object.
}
\description{
Similarly as \code{\link[=write_labels]{write_labels()}}, placing the label in a position ideal to fold
the sheet as an envelope.
This may be usually done for collections of Bryophytes.
}
