% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_labels.R
\name{write_labels}
\alias{write_labels}
\alias{write_labels.specimens}
\alias{write_labels,specimens-method}
\title{Writting Labels for Herbarium Vouchers}
\usage{
write_labels(x, ...)

\method{write_labels}{specimens}(
  x,
  output_file,
  merge = TRUE,
  frame = FALSE,
  date_format = "\%d.\%m.\%Y",
  dim = c(h = 74, w = 105, u = "mm"),
  nup = c(2, 4),
  mar = c(l = 7, r = 7, t = 7, b = 7, u = "mm"),
  classoption = "a4paper",
  ...
)
}
\arguments{
\item{x}{A \linkS4class{specimens} object retrieved by \code{\link[=read_spec]{read_spec()}}.}

\item{...}{Further arguments passed to \code{\link[=list2rmd_doc]{list2rmd_doc()}}. It works only if
\code{'merge = TRUE'}.}

\item{output_file}{A character value indicating the name of the file. It may
include a relative path. This is passed to the homonymous parameter at
\code{\link[=render_rmd]{render_rmd()}}.}

\item{merge}{A logical value indicating whether labels should be merged into
one A4 sheet or remain as individual pages in the output PDF file.}

\item{frame}{A logical value indicating whether a frame should be drawn
around single labels or not. It works only if \code{'merge = TRUE'}.}

\item{date_format}{A character value indicating the format used for the
collection date. It is passed to the function \code{\link[=format.Date]{format.Date()}}.}

\item{dim}{A named vector indicating the dimensions of the single labels. You
need to specify \strong{w} (width), \strong{h} (height), and \strong{u} (units).}

\item{nup}{An integer vector of length 2 indicating the number of labels per
sheet (if \code{'merge = TRUE'}). The first value are number of columns and
the second, number of rows.}

\item{mar}{A named vector indicating the margins. The mandatory names are
\strong{l} (left), \strong{r} (right), \strong{t} (top), \strong{b} (bottom), and
\strong{u} (units).}

\item{classoption}{A character value to be inserted as 'classoption' in the
yaml head for the Rmarkdown document.}
}
\value{
An invisible \linkS4class{rmd_doc} object.
}
\description{
Generating labels for vouchers using exported tables.
}
