% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec.R
\name{read_spec}
\alias{read_spec}
<<<<<<< HEAD
<<<<<<< HEAD
=======
\alias{read_spec.PostgreSQLConnection}
>>>>>>> refs/heads/devel
=======
\alias{read_spec.PostgreSQLConnection}
>>>>>>> refs/heads/devel
\alias{read_spec,PostgreSQLConnection-method}
\title{Import Specimen Information.}
\usage{
read_spec(db, ...)

<<<<<<< HEAD
<<<<<<< HEAD
\S4method{read_spec}{PostgreSQLConnection}(db, adm, bulk, ...)
=======
\method{read_spec}{PostgreSQLConnection}(db, adm, bulk, ...)
>>>>>>> refs/heads/devel
=======
\method{read_spec}{PostgreSQLConnection}(db, adm, bulk, ...)
>>>>>>> refs/heads/devel
}
\arguments{
\item{db}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods (not yet used).}

\item{adm}{Connection to the database containing administrative units as
\linkS4class{PostgreSQLConnection}. If not provided, no administrative units
will be included in the output.}

\item{bulk}{Integer vector including the ID's of the requested bulks
(campaigns or projects).}
}
\value{
An S3 object of class \code{specimens}.
}
\description{
Import function for herbarium specimens stored in own database.
Information is arranged into three tables and formatted as an object of class
\linkS4class{specimens}.

For convenience the function will display the list of updates and prompt a
yes/no/cancel question. Here you should check the printed names.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
