% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec.R
\name{read_spec}
\alias{read_spec}
\alias{read_spec,PostgreSQLConnection,PostgreSQLConnection-method}
\alias{PostgreSQLConnection,PostgreSQLConnection,character-method}
\title{Import Specimen Information.}
\usage{
read_spec(db, adm, ...)

\S4method{read_spec}{PostgreSQLConnection,PostgreSQLConnection}(db, adm, bulk, ...)
}
\arguments{
\item{db, adm}{Connections to the main database and the database containing
administrative units, respectively. Both connections have to be of class
\linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods (not yet used).}

\item{bulk}{Integer vector including the ID's of the requested bulks
(campaigns or projects).}
}
\value{
An S3 object of class \code{specimens}.
}
\description{
Import function for herbarium specimens stored in own database.
Information is arranged into three tables and formatted as an object of class
\linkS4class{specimens}.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
